fix: rename document
